backend/requirements.txt
fastapi[standard]
uvicorn[standard]
pydantic[standard]
pycryptodome[standard]